[update] customer delete confirm

+++ src/main/java/vn/aptech/quanlybanhang/pages/CustomerDeletePage.java
@@ -5,6 +5,7 @@ package vn.aptech.quanlybanhang.pages;
 
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import vn.aptech.quanlybanhang.entities.Customer;
 import vn.aptech.quanlybanhang.entities.Order;
 import vn.aptech.quanlybanhang.service.CustomerService;
 import vn.aptech.quanlybanhang.service.CustomerServiceImpl;
@@ -24,23 +25,32 @@ public class CustomerDeletePage extends Page {
         try {
             CustomerService customerService = new CustomerServiceImpl();
             OrderService orderService = new OrderServiceImpl();
+            Customer customer = null;
             
             int check = AppScanner.scanIntWithMessage(I18n.getEntityMessage("customer", "entity.scan.id.delete"));
             
-            while (customerService.findById(check) == null) {
+            while ((customer = customerService.findById(check)) == null) {
                 I18n.printEntityMessage("customer", "entity.error.idNotFound");
                 check = AppScanner.scanIntWithMessage(I18n.getEntityMessage("customer", "entity.scan.id.delete"));
             }
             
-            Order order = orderService.findByCustomerId(check);
-            if (order == null) {
-                if (customerService.deleteById(check)) {
-                    I18n.printEntityMessage("customer", "entity.msg.deleted");
+            I18n.print(
+                    "entity.msg.foundName", 
+                    I18n.getMessage("customer.label.singular"), 
+                    String.format("%s (%s: %s)", customer.getName(), I18n.getMessage("customer.phone"), customer.getPhone())
+            );
+            
+            if(AppScanner.confirm(I18n.getEntityMessage("customer", "entity.confirm.delete"))) {
+                Order order = orderService.findByCustomerId(check);
+                if (order == null) {
+                    if (customerService.deleteById(check)) {
+                        I18n.printEntityMessage("customer", "entity.msg.deleted");
+                    } else {
+                        I18n.printEntityMessage("customer", "entity.error.deleteFailed");
+                    }
                 } else {
-                    I18n.printEntityMessage("customer", "entity.error.deleteFailed");
+                    I18n.print("customer.error.cannotDelete");
                 }
-            } else {
-                I18n.print("custom.error.cannotDelete");
             }
 
         } catch (Exception ex) {
